+++ bin/faucet/src/handlers.rs
@@ -1,60 +1,71 @@
-use actix_web::{get, http::header, post, web, HttpResponse, Result};
+use axum::{
+    extract::{Path, State},
+    http::{Response, StatusCode},
+    response::IntoResponse,
+    Json,
+};
+use http::header;
+use http_body_util::Full;
 use miden_objects::{
     accounts::AccountId,
     notes::{NoteDetails, NoteExecutionMode, NoteFile, NoteId, NoteTag},
     utils::serde::Serializable,
 };
 use serde::{Deserialize, Serialize};
+use tonic::body;
 use tracing::info;
 
-use crate::{errors::FaucetError, state::FaucetState};
+use crate::{errors::ProcessError, state::FaucetState, COMPONENT};
 
 #[derive(Deserialize)]
-struct FaucetRequest {
+pub struct FaucetRequest {
     account_id: String,
     is_private_note: bool,
     asset_amount: u64,
 }
 
 #[derive(Serialize)]
-struct FaucetMetadataReponse {
+pub struct FaucetMetadataReponse {
     id: String,
     asset_amount_options: Vec<u64>,
 }
 
-#[get("/get_metadata")]
-pub async fn get_metadata(state: web::Data<FaucetState>) -> HttpResponse {
+pub async fn get_metadata(
+    State(state): State<FaucetState>,
+) -> (StatusCode, Json<FaucetMetadataReponse>) {
     let response = FaucetMetadataReponse {
         id: state.id.to_string(),
         asset_amount_options: state.config.asset_amount_options.clone(),
     };
 
-    HttpResponse::Ok().json(response)
+    (StatusCode::OK, Json(response))
 }
 
-#[post("/get_tokens")]
 pub async fn get_tokens(
-    req: web::Json<FaucetRequest>,
-    state: web::Data<FaucetState>,
-) -> Result<HttpResponse> {
+    State(state): State<FaucetState>,
+    Json(req): Json<FaucetRequest>,
+) -> Result<impl IntoResponse, ProcessError> {
     info!(
-        "Received a request with account_id: {}, is_private_note: {}, asset_amount: {}",
-        req.account_id, req.is_private_note, req.asset_amount
+        target: COMPONENT,
+        account_id = %req.account_id,
+        is_private_note = %req.is_private_note,
+        asset_amount = %req.asset_amount,
+        "Received a request",
     );
 
     // Check that the amount is in the asset amount options
     if !state.config.asset_amount_options.contains(&req.asset_amount) {
-        return Err(FaucetError::BadRequest("Invalid asset amount.".to_string()).into());
+        return Err(ProcessError::BadRequest("Invalid asset amount".to_string()));
     }
 
     let mut client = state.client.lock().await;
 
     // Receive and hex user account id
     let target_account_id = AccountId::from_hex(req.account_id.as_str())
-        .map_err(|err| FaucetError::BadRequest(err.to_string()))?;
+        .map_err(|err| ProcessError::BadRequest(err.to_string()))?;
 
     // Execute transaction
-    info!("Executing mint transaction for account.");
+    info!(target: COMPONENT, "Executing mint transaction for account.");
     let (executed_tx, created_note) = client.execute_mint_transaction(
         target_account_id,
         req.is_private_note,
@@ -62,7 +73,7 @@ pub async fn get_tokens(
     )?;
 
     // Run transaction prover & send transaction to node
-    info!("Proving and submitting transaction.");
+    info!(target: COMPONENT, "Proving and submitting transaction.");
     let block_height = client.prove_and_submit_transaction(executed_tx).await?;
 
     let note_id: NoteId = created_note.id();
@@ -80,17 +91,33 @@ pub async fn get_tokens(
     }
     .to_bytes();
 
-    info!("A new note has been created: {}", note_id);
+    info!(target: COMPONENT, %note_id, "A new note has been created");
 
     // Send generated note to user
-    Ok(HttpResponse::Ok()
-        .content_type("application/octet-stream")
-        .append_header(header::ContentDisposition {
-            disposition: actix_web::http::header::DispositionType::Attachment,
-            parameters: vec![actix_web::http::header::DispositionParam::Filename(
-                "note.mno".to_string(),
-            )],
-        })
-        .append_header(("Note-Id", note_id.to_string()))
-        .body(bytes))
+    Response::builder()
+        .status(StatusCode::OK)
+        .header(header::CONTENT_TYPE, "application/octet-stream")
+        .header(header::CONTENT_DISPOSITION, "attachment; filename=note.mno")
+        .header("Note-Id", note_id.to_string())
+        .body(body::boxed(Full::from(bytes)))
+        .map_err(|err| ProcessError::InternalServerError(err.to_string()))
+}
+
+pub async fn get_index(state: State<FaucetState>) -> Result<impl IntoResponse, ProcessError> {
+    get_static_file(state, Path("index.html".to_string())).await
+}
+
+pub async fn get_static_file(
+    State(state): State<FaucetState>,
+    Path(path): Path<String>,
+) -> Result<impl IntoResponse, ProcessError> {
+    info!(target: COMPONENT, path, "Serving static file");
+
+    let static_file = state.static_files.get(path.as_str()).ok_or(ProcessError::NotFound(path))?;
+
+    Response::builder()
+        .status(StatusCode::OK)
+        .header(header::CONTENT_TYPE, static_file.mime_type)
+        .body(body::boxed(Full::from(static_file.data)))
+        .map_err(|err| ProcessError::InternalServerError(err.to_string()))
 }